azure-website-summarizer/requirements.txt
azure-ai-textanalytics
beautifulsoup4
requests
flask
python-dotenv
streamlit
azure-identity
azure-keyvault-secrets
openai
